Handle edge case
The regex up to now could not handle a quoted string inside another, i.e. ('"lala"').

If that string had some % symbol inside ('"%lala"') it already though that the string
was finished and thus the % symbol was the one triggering a %-format string.

Now we check for a single quote or double quoted string.